reverted cores to int

+++ service/cluster/searcher/config/cpu.go
@@ -2,7 +2,7 @@ package config
 
 // CPU configures the machine CPU.
 type CPU struct {
-	Cores float64 `json:"cores"`
+	Cores int `json:"cores"`
 }
 
 // DefaultCPU provides a default CPU configuration by best effort.
